view restructure
